Remove "international: True" X resource.
Setting the widget resource "international" to true changes how the
application handles fonts in an incompatible manner.

The "international" resource also causes the application to emit the
message "Warning: Missing charsets in String to FontSet conversion"
unless the LANG or LC_ALL environment variable is set to "C".

The setting being reverted here was added in xlbiff 4.2.

Users for whom "international" was useful, or who had already updated
their resources to handle it, can get the 4.2 behavior back by adding
the following line to an X Resources file or -xrm argument:

*international: true