Change default mail path to /var/mail
The directory containing user mailbox files has been standardized as
/var/mail since Filesystem Hierarchy Standard version 2.0, published
in 1997.  All Linux and BSD distributions examined use /var/mail.

Change the name of the variable used, from MAILPATH to XLBIFF_MAILPATH,
so as not to conflict with some shells that use MAILPATH and leak the
shell environment variable into the build environment.