Add a test that scanCommand works out of the box
New script test-scan-interface.sh tests the way xlbiff retries
the "scan" command.

This test is not run automatically by the build, because it requires
nmh and xvfb to be installed; these are not otherwise build
dependencies.  On a system with the test dependencies installed, the
test can be run manually.

This test is referred to by a Debian-specific autopkgtest control
file.  (autopkgtest can install dependencies needed for the test.)