mailer() action always re-checks for mail
Even with no mailerCommand, the mailer() action checks for mail.
This makes the action useful even without defining a mailerCommand.

Change the order of Popup and checking for mail in the mailer() action.
Previously, xlbiff could pop up only to discover there was no mail and
pop down again immediately.

In the manual, move the discussion of the mailer() action to the section
with the other actions.