Refactor Popup/lbiffRealize and Popdown/lbiffUnrealize
Previously lbiffUnrealize would sometimes call Popdown and sometimes
unrealize the widget.  But the widget always needs to be unrealized
(so that the its position gets reset when it pops up again), and the
bookkeeping done in Popdown always needs to be done.  When Popdown was
not called, the refresh time and ledPopdown were not honored.  Now
lbiffUnrealize always unrealizes and updates internals, and Popdown is
a wrapper around lbiffUnrealize.

lbiffRealize continues to call Popup.  Handling the "bottom" behavior
is moved to Popup because it always needs to be done.  If "bottom" is
not handled by Popup, the window drifts off the bottom when
mailerCommand does not change the mailbox.