Improve robustness of test function get_window_corners
Ensure get_window_corners does not write to stderr, even if it fails
to find the window.  Do not use stdout for return value, so verbose
logging does not interfere with the value.

Replace test function util_verb with util_logv; writing everything to
the log file makes debugging easier.